aggiunti campi statistiche ad entità progetto
